requirements.txt
Package                   Version
------------------------- -----------
agate                     1.9.1
altair                    5.5.0
annotated-types           0.7.0
async-timeout             5.0.1
attrs                     25.3.0
babel                     2.17.0
blinker                   1.9.0
cachetools                6.1.0
certifi                   2025.6.15
charset-normalizer        3.4.2
click                     8.2.1
colorama                  0.4.6
daff                      1.4.2
dbt-adapters              1.16.0
dbt-common                1.25.0
dbt-core                  1.10.2
dbt-extractor             0.6.0
dbt-postgres              1.9.0
dbt-protos                1.0.330
dbt-semantic-interfaces   0.8.4
deepdiff                  7.0.1
Faker                     37.4.0
gitdb                     4.0.12
GitPython                 3.1.44
greenlet                  3.2.3
idna                      3.10
importlib_metadata        8.7.0
isodate                   0.6.1
Jinja2                    3.1.6
jsonschema                4.24.0
jsonschema-specifications 2025.4.1
leather                   0.4.0
MarkupSafe                3.0.2
mashumaro                 3.14
more-itertools            10.7.0
msgpack                   1.1.1
narwhals                  1.44.0
networkx                  3.4.2
numpy                     2.2.6
ordered-set               4.1.0
packaging                 25.0
pandas                    2.3.0
parsedatetime             2.6
pathspec                  0.12.1
pillow                    11.2.1
pip                       22.0.2
protobuf                  5.29.5
psycopg2-binary           2.9.10
pyarrow                   20.0.0
pydantic                  2.11.7
pydantic_core             2.33.2
pydeck                    0.9.1
python-dateutil           2.9.0.post0
python-dotenv             1.1.0
python-slugify            8.0.4
pytimeparse               1.1.8
pytz                      2025.2
PyYAML                    6.0.2
redis                     6.2.0
referencing               0.36.2
requests                  2.32.4
rpds-py                   0.25.1
setuptools                59.6.0
six                       1.17.0
smmap                     5.0.2
snowplow-tracker          1.1.0
spotipy                   2.25.1
SQLAlchemy                2.0.41
sqlparse                  0.5.3
streamlit                 1.46.0
tenacity                  9.1.2
text-unidecode            1.3
toml                      0.10.2
tornado                   6.5.1
typing_extensions         4.14.0
typing-inspection         0.4.1
tzdata                    2025.2
urllib3                   2.5.0
watchdog                  6.0.0
zipp                      3.23.0
